Create rr.c
This is a rusian roulette game which is seeded. 
It xors a random CPU word(32 bits) by 7 at the time to simulate corruption. Due to the von neuman architecture it will eventually corrupt its own memory subsequently killing its own task . The current seed loaded up will put the CPU into an infinite loop which it will do 2 things which are visible in VGA screen.